Module oak_crypto::encryption_key

Structs

EncryptionKey

Traits

AsyncEncryptionKeyHandle

EncryptionKeyHandle

Exposes the ability to derive a session key from the provided encapsulated private key, using a private key that has been endorsed in the Attestation Evidence.

Functions

generate_encryption_key_pair

Generates a random encryption key pair and returns an instance of the EncryptionKey and a NIST P-256 SEC1 encoded point public key. https://secg.org/sec1-v2.pdf