Trait oak_crypto::encryption_key::EncryptionKeyHandle

pub trait EncryptionKeyHandle {
    // Required method
    fn generate_recipient_context(
        &self,
        encapsulated_public_key: &[u8],
    ) -> Result<RecipientContext>;
}

Exposes the ability to derive a session key from the provided encapsulated private key, using a private key that has been endorsed in the Attestation Evidence.

Required Methods

fn generate_recipient_context( &self, encapsulated_public_key: &[u8], ) -> Result<RecipientContext>